Refactor : ModalBackground로 모달 형식 통일
