+++ src/Pages/AdminPage/AdminContainer/AdminModal/ModalAddVacationAll.jsx
@@ -1,8 +1,7 @@
 import { useState } from 'react';
+import { ModalBackground } from 'Components';
 import { VACATION } from 'Utils/constants';
 import { VacationService } from 'API';
-
-import styled from 'styled-components';
 
 import Button from '@mui/material/Button';
 import Slider from '@mui/material/Slider';
@@ -55,7 +54,7 @@ const ModalAddVacationAll = ({
     setValue(event.target.value);
   };
   return (
-    <ModalAddVacationAllBody>
+    <ModalBackground setIsOpen={setIsOpen}>
       <h1>일괄 휴가 변경</h1>
       <h3>참가 중인 모든 사용자들의 휴가를 증가하거나 감소합니다.</h3>
       <h3>얼마나 변경하시겠습니까?</h3>
@@ -70,19 +69,12 @@ const ModalAddVacationAll = ({
         min={-0.5}
         max={3}
       />
-      <Button onClick={() => handleCloseModal(true)}>확인</Button>
-      <Button onClick={() => handleCloseModal(false)}>취소</Button>
-    </ModalAddVacationAllBody>
+      <div className="buttons">
+        <Button onClick={() => handleCloseModal(true)}>확인</Button>
+        <Button onClick={() => handleCloseModal(false)}>취소</Button>
+      </div>
+    </ModalBackground>
   );
 };
-
-const ModalAddVacationAllBody = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(216, 216, 216, 0.9);
-`;
 
 export default ModalAddVacationAll;
